Autocleanup

+++ src/main/java/xyz/brassgoggledcoders/workshop/renderer/ChalkWritingTileEntityRenderer.java
@@ -1,22 +1,16 @@
 package xyz.brassgoggledcoders.workshop.renderer;
 
 import com.mojang.blaze3d.matrix.MatrixStack;
-import com.mojang.blaze3d.vertex.IVertexBuilder;
-import net.minecraft.block.*;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.WallSignBlock;
 import net.minecraft.client.gui.FontRenderer;
 import net.minecraft.client.gui.RenderComponentsUtil;
-import net.minecraft.client.renderer.Atlases;
 import net.minecraft.client.renderer.IRenderTypeBuffer;
-import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.Vector3f;
-import net.minecraft.client.renderer.model.Material;
-import net.minecraft.client.renderer.model.Model;
-import net.minecraft.client.renderer.model.ModelRenderer;
 import net.minecraft.client.renderer.texture.NativeImage;
 import net.minecraft.client.renderer.tileentity.TileEntityRenderer;
 import net.minecraft.client.renderer.tileentity.TileEntityRendererDispatcher;
 import net.minecraft.item.DyeColor;
-import net.minecraft.tileentity.SignTileEntity;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
@@ -40,22 +34,22 @@ public class ChalkWritingTileEntityRenderer extends TileEntityRenderer<ChalkWrit
         matrixStackIn.translate(0.0D, -0.3125D, -0.4375D);
 
         FontRenderer fontrenderer = this.renderDispatcher.getFontRenderer();
-        matrixStackIn.translate(0.0D, (double)0.33333334F, (double)0.046666667F);
+        matrixStackIn.translate(0.0D, 0.33333334F, 0.046666667F);
         matrixStackIn.scale(0.010416667F, -0.010416667F, 0.010416667F);
         int i = DyeColor.WHITE.getTextColor();
-        int j = (int)((double) NativeImage.getRed(i) * 0.4D);
-        int k = (int)((double)NativeImage.getGreen(i) * 0.4D);
-        int l = (int)((double)NativeImage.getBlue(i) * 0.4D);
+        int j = (int) ((double) NativeImage.getRed(i) * 0.4D);
+        int k = (int) ((double) NativeImage.getGreen(i) * 0.4D);
+        int l = (int) ((double) NativeImage.getBlue(i) * 0.4D);
         int i1 = NativeImage.getCombined(0, l, k, j);
 
-        for(int j1 = 0; j1 < 4; ++j1) {
+        for (int j1 = 0; j1 < 4; ++j1) {
             String s = tileEntityIn.getRenderText(j1, (p_212491_1_) -> {
                 List<ITextComponent> list = RenderComponentsUtil.splitText(p_212491_1_, 90, fontrenderer, false, true);
                 return list.isEmpty() ? "" : list.get(0).getFormattedText();
             });
             if (s != null) {
-                float f3 = (float)(-fontrenderer.getStringWidth(s) / 2);
-                fontrenderer.renderString(s, f3, (float)(j1 * 10 - tileEntityIn.signText.length * 5), i1, false, matrixStackIn.getLast().getMatrix(), bufferIn, false, 0, combinedLightIn);
+                float f3 = (float) (-fontrenderer.getStringWidth(s) / 2);
+                fontrenderer.renderString(s, f3, (float) (j1 * 10 - tileEntityIn.signText.length * 5), i1, false, matrixStackIn.getLast().getMatrix(), bufferIn, false, 0, combinedLightIn);
             }
         }
